% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{automatedRstoxTest}
\alias{automatedRstoxTest}
\title{Function for running all test projects and comparing outputs with previous outputs.}
\usage{
automatedRstoxTest(root = list(windows = "\\\\\\\\delphi", unix =
  "/Volumes"), path = "pc_prog/S2D/stox/StoXVerTest",
  copyFromServer = TRUE, process = c("run", "diff"),
  diffs = c("Rdata", "images", "text", "baseline"), projectInd = NULL,
  nlines = 50, mem.size = 1.6e+10, nwarnings = 10000, n = 1L,
  skipError = FALSE)
}
\arguments{
\item{root}{A list of specifyers for the root directory to the central server.}

\item{path}{The relative path from the root.}

\item{copyFromServer}{Logical: If TRUE, copy the latest original projects, outputs and diffs in the server to the local directory.}

\item{process}{Which steps to run in the testing used mostly to reduce processing time for development and bug fixing.}

\item{diffs}{Which diffs to include, also used to reduce processing time.}

\item{nlines}{The number of lines to display for diffs between text files.}
}
\description{
Function for running all test projects and comparing outputs with previous outputs.
}
\keyword{internal}
