% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_Rstox_package}
\alias{build_Rstox_package}
\alias{package_specs}
\title{Function for building Rstox packages}
\usage{
build_Rstox_package(packageName, version = "1.0", Rversion = "3.5",
  pckversion = NULL, suggests = NULL,
  githubRoot = "https://github.com/Sea2Data", onCran = FALSE,
  license = "LGPL-3", rootDir = NULL, title = NULL,
  description = NULL, details = NULL, .onLoad = NULL,
  .onAttach = NULL, misc = NULL, authors = NULL, check = FALSE)

package_specs(packageName, version = "1.0", Rversion = "3.5",
  pckversion = NULL, suggests = NULL,
  githubRoot = "https://github.com/Sea2Data", onCran = FALSE,
  license = "LGPL-3", rootDir = NULL, title = NULL,
  description = NULL, details = NULL, authors = NULL,
  .onLoad = NULL, .onAttach = NULL, misc = NULL)
}
\arguments{
\item{check}{Logical: If TRUE run devtools::check() on the package.}

\item{dir}{The directory holding the package structure.}
}
\description{
\code{build_Rstox_package} is used in the continous development of Rstox, writing .onLoad, .onAttach, pkgname, DESCRIPTION and README files and adding dependencies to the NAMEPACE file.\cr \cr
\code{package_specs} gets all specifications of the package from separate funcitons for each package into a list.\cr \cr
}
