% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildRstoxPackage}
\alias{buildRstoxPackage}
\alias{packageSpecs}
\title{Function for building Rstox packages}
\usage{
buildRstoxPackage(
  packageName,
  accountName = "StoXProject",
  version = "current",
  Rversion = "3.5",
  imports = NULL,
  suggests = NULL,
  linkingto = NULL,
  importToNamespace = NULL,
  internal.dependencies = NULL,
  internal.suggests = NULL,
  additional_repositories = NULL,
  onCran = FALSE,
  license = "LGPL-3",
  rootDir = NULL,
  title = NULL,
  description = NULL,
  details = NULL,
  .onLoad = NULL,
  .onUnload = NULL,
  .onAttach = NULL,
  misc = NULL,
  authors = NULL,
  check = FALSE,
  noRcpp = FALSE,
  addManual = FALSE,
  addIndividualManuals = FALSE,
  globalVariables = NULL,
  type = c("patch", "minor", "major"),
  prerelease = FALSE,
  date = NULL,
  avoid_compileAttributes_error = FALSE,
  VignetteBuilder = NULL
)

packageSpecs(
  packageName,
  accountName = "StoXProject",
  version = "current",
  Rversion = "3.5",
  imports = NULL,
  suggests = NULL,
  linkingto = NULL,
  importToNamespace = importToNamespace,
  internal.dependencies = NULL,
  internal.suggests = NULL,
  additional_repositories = NULL,
  onCran = FALSE,
  license = "LGPL-3",
  rootDir = NULL,
  title = NULL,
  description = NULL,
  details = NULL,
  authors = NULL,
  .onLoad = NULL,
  .onUnload = NULL,
  .onAttach = NULL,
  misc = NULL,
  type = c("patch", "minor", "major"),
  prerelease = FALSE,
  date = NULL,
  avoid_compileAttributes_error = FALSE,
  VignetteBuilder = NULL
)
}
\arguments{
\item{check}{Logical: If TRUE run devtools::check() on the package.\code{\link[RstoxBase]{StratumArea}}}

\item{dir}{The directory holding the package structure.}
}
\description{
\code{buildRstoxPackage} is used in the continous development of Rstox, writing .onLoad, .onAttach, pkgname, DESCRIPTION and README files and adding dependencies to the NAMEPACE file.\cr \cr
\code{packageSpecs} gets all specifications of the package from separate funcitons for each package into a list.\cr \cr
}
