% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{copyCurrentToServer}
\alias{copyCurrentToServer}
\title{Copy the test run to the central srever.}
\usage{
copyCurrentToServer(root = list(windows = "\\\\\\\\delphi", unix =
  "/Volumes"), path = "pc_prog/S2D/stox/StoXVerTest",
  toCopy = c("Diff", "Output", "ProjOrig"), overwrite = FALSE,
  msg = FALSE, n = 1)
}
\arguments{
\item{root}{A list of specifyers for the root directory to the central server.}

\item{path}{The relative path from the root.}

\item{toCopy}{The sub folders to copy files from.}

\item{overwrite}{Logical: If TRUE, overwrite the files on the central directory.}

\item{msg}{Logical: If TRUE, print progress to the console.}

\item{n}{The number of runs (one runfor each version tested) to copy.}
}
\description{
Copy the test run to the central srever.
}
\keyword{internal}
