% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{getLatestDir}
\alias{getLatestDir}
\title{Get the latest directory of}
\usage{
getLatestDir(dir, op = "<", n = 1)
}
\arguments{
\item{dir}{The directory of subdirectories named with the Rstox and stox-lib versions.}

\item{op}{The operator used to get the latest. Do not mess with this unless you know what you are doing.}

\item{n}{The number of directories to return, as used in utils::tail().}
}
\description{
Get the latest directory of
}
\keyword{internal}
