% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{getPlatformID}
\alias{getPlatformID}
\title{Get the platform ID of an operating system}
\usage{
getPlatformID()
}
\arguments{
\item{var}{The element of Sys.info() used as identifyer of the platform.}
}
\description{
Get the platform ID of an operating system
}
\keyword{internal}
