% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{getServerPath}
\alias{getServerPath}
\title{Get the path to the server, depending on the local platform (Mac, Windows).}
\usage{
getServerPath(root = list(windows = "\\\\\\\\delphi", unix = "/Volumes"),
  path = "pc_prog/S2D/stox/StoXVerTest")
}
\arguments{
\item{root}{A list of specifyers for the root directory to the central server.}

\item{path}{The relative path from the root.}
}
\description{
Get the path to the server, depending on the local platform (Mac, Windows).
}
\keyword{internal}
