% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{getTestFolderStructure}
\alias{getTestFolderStructure}
\title{Small funciton to get a list of the file paths to the sub folders of the test directory (ProjOrig, Proj, Output, Diff)}
\usage{
getTestFolderStructure(x)
}
\arguments{
\item{x}{The test directory}
}
\description{
Small funciton to get a list of the file paths to the sub folders of the test directory (ProjOrig, Proj, Output, Diff)
}
