% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{prepareStoX}
\alias{prepareStoX}
\title{Function for building Rstox packages}
\usage{
prepareStoX(
  accountName = "StoXProject",
  version = "current",
  date = NULL,
  rootDir = NULL,
  type = c("patch", "minor", "major")
)
}
\arguments{
\item{dir}{The directory holding the package structure.}

\item{check}{Logical: If TRUE run devtools::check() on the package.\code{\link[RstoxBase]{StratumArea}}}
}
\description{
\code{buildRstoxPackage} is used in the continous development of Rstox, writing .onLoad, .onAttach, pkgname, DESCRIPTION and README files and adding dependencies to the NAMEPACE file.\cr \cr
\code{packageSpecs} gets all specifications of the package from separate funcitons for each package into a list.\cr \cr
}
