% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RstoxBuild-package}
\alias{RstoxBuild}
\alias{RstoxBuild-package}
\title{Package for building all Rstox packages
 This package contains functionality for building the Rstox packages (Rstox, RstoxFramework, RstoxData, RstoxECA, RstoxSurveyPlanner, RstoxTempdoc, and even RstoxBuild). The function defines titles, descriptions, dependencies, authors, install instructions and other info for all the packages.
 This package is used for building the Rstox packages. All changes to authors, descriptions, suggests and other outputs of the function \code{package_specs} should be changed in this package, and not in the individual packages. #'}
\description{
This package contains functionality for building the Rstox packages (Rstox, RstoxFramework, RstoxData, RstoxECA, RstoxSurveyPlanner, RstoxTempdoc, and even RstoxBuild). The function defines titles, descriptions, dependencies, authors, install instructions and other info for all the packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IMR-StoX/RstoxBuild}
  \item Report bugs at \url{https://github.com/IMR-StoX/RstoxBuild/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

}
