% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RstoxBuild}
\alias{RstoxBuild}
\alias{RstoxBuild-package}
\title{Package for Building All Rstox Packages}
\description{
Building the Rstox packages (Rstox, RstoxFramework, RstoxData, RstoxFDA, RstoxSurveyPlanner, RstoxTempdoc, and even RstoxBuild), and semi-automated testing of Rstox though test projects.
}
\details{
The package defines titles, descriptions, dependencies, authors, install instructions and other info for all the packages. All changes to authors, descriptions, suggests and other outputs of the function \code{packageSpecs} should be changed in this package, and not in the individual packages. The package also contains functionality for semi-automated testing of Rstox on a set of test projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxBuild}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxBuild/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

}
