% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoXVerTest.R
\name{runProject}
\alias{runProject}
\title{Function for running the r scripts of a project and copying the relevant output files to the "Output" directory:}
\usage{
runProject(projectName, progressFile, outputDir, ind = NULL)
}
\arguments{
\item{projectName}{The path to the project.}

\item{progressFile}{The paht to the progress file.}

\item{outputDir}{The path to the directory in which to put the output files to be compared for the project (everything in "output" except trash files, baseline and baseline report output, and the project.xml file).}
}
\description{
Function for running the r scripts of a project and copying the relevant output files to the "Output" directory:
}
\keyword{internal}
