% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildRstoxPackage}
\alias{buildRstoxPackage}
\alias{packageSpecs}
\title{Function for building Rstox packages}
\usage{
buildRstoxPackage(packageName, version = "1.0", Rversion = "3.5",
  pckversion = NULL, suggests = NULL,
  githubRoot = "https://github.com/StoXProject", onCran = FALSE,
  license = "LGPL-3", rootDir = NULL, title = NULL,
  description = NULL, details = NULL, .onLoad = NULL,
  .onAttach = NULL, misc = NULL, authors = NULL, check = FALSE,
  noRcpp = FALSE)

packageSpecs(packageName, version = "1.0", Rversion = "3.5",
  pckversion = NULL, suggests = NULL,
  githubRoot = "https://github.com/StoXProject", onCran = FALSE,
  license = "LGPL-3", rootDir = NULL, title = NULL,
  description = NULL, details = NULL, authors = NULL,
  .onLoad = NULL, .onAttach = NULL, misc = NULL)
}
\arguments{
\item{check}{Logical: If TRUE run devtools::check() on the package.}

\item{dir}{The directory holding the package structure.}
}
\description{
\code{buildRstoxPackage} is used in the continous development of Rstox, writing .onLoad, .onAttach, pkgname, DESCRIPTION and README files and adding dependencies to the NAMEPACE file.\cr \cr
\code{packageSpecs} gets all specifications of the package from separate funcitons for each package into a list.\cr \cr
}
