% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{getDESCRIPTION}
\alias{getDESCRIPTION}
\alias{getREADME}
\alias{addImportsToDESCRIPTION}
\alias{getPkgname}
\title{Utility functions for getting and adding imports, DESCRIPTION, README and kgname to the package.}
\usage{
getDESCRIPTION(spec)

getREADME(spec)

addImportsToDESCRIPTION(spec, cpp = FALSE)

getPkgname(spec)
}
\arguments{
\item{spec}{A list of package specifications returned from \code{\link{packageSpecs}}.}
}
\description{
\code{getDESCRIPTION} gets the DESCRIPTION text to write to the DESCRIPTION file, adding authors, R-dependency, title, description and other info stored in the input \code{spec}. Note that the package imports are not added here, but later in the \code{addImportsToDESCRIPTION} function.\cr \cr
\code{getREADME} gets the README file in the Rstox style, which has package and R version in the first two lines, followed by description, install instructions, miscellaneous info, and release notes retrieved from the NEWS file.\cr \cr
\code{addImportsToDESCRIPTION} adds the imports to the DESCRIPTION file.\cr \cr
\code{getPkgname} gets the pkgname text to write to the pkgname.R file.
}
